Se cambian los nombres de los modal.
Se comienza a extraer la logica de axios a clases aparte
