+++ src/components/Modal/CreateCareer.js
@@ -2,59 +2,66 @@ import React, { Component } from "react";
 import swal from "sweetalert";
 import axios from "axios";
 import $ from "jquery";
-import Validator from "../../helpers/Validations";
 import { handleSimpleInputChange } from "../../helpers/Handles";
+import Validator from "../../helpers/Validations";
 
 /**
  * * Componente que muestra la ventana y elementos correspondientes
- * * para la edición de una carrera universitaria
+ * * para la creación de una nueva carrera
  */
-export default class ModalCareerEdit extends Component {
+export default class CreateCareer extends Component {
   constructor(props) {
     super(props);
     this.state = {
       name: "",
+      career_code: "",
       degree: "",
     };
 
     //bind
-    this.validateShow = this.validateShow.bind(this);
+    this.show = this.show.bind(this);
     this.handleChange = handleSimpleInputChange.bind(this);
     this.handleSubmit = this.handleSubmit.bind(this);
 
     //ref
+    this.careerCodeError = React.createRef();
+    this.careerDegreeError = React.createRef();
     this.careerNameError = React.createRef();
   }
 
   /**
-   * * Función que valida si el componente debe mostrarse, dependiendo
-   * * de las propiedades que le entran por parámetro. En este caso el
-   * * código de la carrera universitaria debe estar definido.
+   * * Función que muestra el componente y limpia las variables
+   * * del estado, así como los mensajes de error correspondientes
    */
-  validateShow() {
-    if (this.props.career_code !== "") {
-      this.setState({
-        name: this.props.career_name,
-        degree: this.props.career_degree,
-      });
-      this.careerNameError.current.style.display = "none";
-      $("#modalCareerEdit").modal("toggle");
-    } else {
-      swal(
-        "¡Atención!",
-        "Debe seleccionar una carrera de la lista.",
-        "warning"
-      );
-    }
+  show() {
+    this.setState({ name: "", career_code: "" });
+
+    this.careerCodeError.current.style.display = "none";
+    this.careerDegreeError.current.style.display = "none";
+    this.careerNameError.current.style.display = "none";
+    $("#modalCareer").modal("toggle");
   }
 
   /**
    * * Función que maneja el envío del formulario.
-   * * Se encarga de editar la carrera universitaria si
-   * * no se presentan errores en el nombre y grado seleccionado.
+   * * Se encarga de crear la carrera universitaria si
+   * * no se presentan errores en el nombre, código y grado seleccionado.
    */
   async handleSubmit(event) {
     event.preventDefault();
+    
+    const codeError = Validator.validateSimpleText(
+      this.state.career_code,
+      this.careerCodeError.current,
+      20,
+      "onlyNumber"
+    );
+
+    const degreeError = Validator.validateSimpleSelect(
+      this.state.degree,
+      this.careerDegreeError.current
+    );
+
     const nameError = Validator.validateSimpleText(
       this.state.name,
       this.careerNameError.current,
@@ -62,55 +69,81 @@ export default class ModalCareerEdit extends Component {
       "textSpecial"
     );
 
-    if (!nameError) {
+    if (!nameError && !codeError && !degreeError) {
       const career = {
         name: this.state.name,
+        career_code: this.state.career_code,
         degree: this.state.degree,
       };
-      await axios.put(`/career/` + this.props.career_code, career);
-      this.props.getCareer();
-      $("#modalCareerEdit").modal("hide");
-      swal("¡Listo!", "Se editó la carrera exitosamente.", "success");
-
-      this.props.refreshThis({
-        career_code: "",
-        career_key: this.props.select_key + 1,
+      const exist = await axios.post("/career_exists", {
+        id: career.career_code,
       });
+      if (!exist.data.careerexists) {
+        await axios.post(`/career`, career);
+        this.props.getCareers();
+        $("#modalCareer").modal("hide");
+        swal("¡Listo!", "Se creó la nueva carrera exitosamente.", "success");
+      } else {
+        swal(
+          "¡Atención!",
+          "No se creó la carrera debido a que su código ya se encuentra asociado",
+          "warning"
+        );
+      }
     }
   }
 
-  // Función que renderiza el componente para mostrarlo en pantalla
   render() {
     return (
       <div className="modal-container">
         <button
           type="button"
           className="btn btn-primary btn-md"
-          data-target="#modalCareerEdit"
-          onClick={this.validateShow}
+          data-target="#modalCareer"
+          onClick={this.show}
+          disabled={
+            this.props.parent === "ver" || this.props.parent === "registro"
+              ? this.props.disabled
+              : ""
+          }
         >
-          <i className="fas fa-edit"></i>
+          <i className="fas fa-plus"></i>
         </button>
-        <div className="modal fade" id="modalCareerEdit" role="dialog">
+        <div className="modal fade" id="modalCareer" role="dialog">
           <div className="modal-dialog modal-md modal-dialog-centered">
             <div className="modal-content">
               <div className="modal-header">
-                <h4 className="modal-title">Editar carrera</h4>
+                <h4 className="modal-title">Crear nueva carrera</h4>
                 <button type="button" className="close" data-dismiss="modal">
                   &times;
                 </button>
               </div>
               <div className="modal-body">
                 <div className="form-group">
+                  <label htmlFor="careerCode">Código de la carrera</label>
+                  <input
+                    className="form-control"
+                    type="text"
+                    id="careerCode"
+                    name="career_code"
+                    value={this.state.career_code}
+                    onChange={this.handleChange}
+                  ></input>
+                  <div
+                    className="alert alert-danger"
+                    style={{ fontSize: 12 }}
+                    ref={this.careerCodeError}
+                  ></div>
+                </div>
+                <div className="form-group">
                   <label htmlFor="degree">Grado académico</label>
                   <select
                     className="form-control"
                     id="degree"
                     name="degree"
-                    value={this.state.degree}
                     onChange={this.handleChange}
                   >
-                    <option className="select-cs" value="" defaultValue>
+                    <option className="select-cs" value="default" defaultValue>
                       Seleccione un grado
                     </option>
                     <option value="Diplomado">Diplomado</option>
@@ -119,13 +152,17 @@ export default class ModalCareerEdit extends Component {
                     <option value="Maestría">Maestría</option>
                     <option value="Doctorado">Doctorado</option>
                   </select>
+                  <div
+                    className="alert alert-danger"
+                    style={{ fontSize: 12 }}
+                    ref={this.careerDegreeError}
+                  ></div>
                 </div>
                 <div className="form-group">
-                  <label htmlFor="nombreCareer">Nombre</label>
+                  <label htmlFor="name">Nombre</label>
                   <input
                     className="form-control"
                     type="text"
-                    id="nombreCareer"
                     name="name"
                     value={this.state.name}
                     onChange={this.handleChange}
@@ -142,7 +179,7 @@ export default class ModalCareerEdit extends Component {
                   Cancelar
                 </button>
                 <button className="btn btn-primary" onClick={this.handleSubmit}>
-                  Guardar
+                  Crear
                 </button>
               </div>
             </div>
