+++ src/components/Modal/EditAsso.js
@@ -5,44 +5,40 @@ import $ from "jquery";
 import Validator from "../../helpers/Validations";
 import { handleSimpleInputChange } from "../../helpers/Handles";
 
-/*
-    
-    
-*/
 /**
  * * Componente que muestra la ventana y elementos correspondientes
- * * para la creación de una nueva carrera asociada
+ * * para la edición de una carrera asociada
  */
-export default class ModalAsso extends Component {
+export default class EditAsso extends Component {
   constructor(props) {
     super(props);
     this.state = {
       name: "",
     };
 
-    //Bind
+    //bind
     this.validateShow = this.validateShow.bind(this);
     this.handleChange = handleSimpleInputChange.bind(this);
     this.handleSubmit = this.handleSubmit.bind(this);
 
     //ref
-    this.assoNameError = React.createRef();
+    this.assoEditNameError = React.createRef();
   }
 
   /**
    * * Función que valida si el componente debe mostrarse, dependiendo
    * * de las propiedades que le entran por parámetro. En este caso el
-   * * id del centro educativo debe estar definido.
+   * * id de la carrera asociada debe estar definido.
    */
   validateShow() {
-    if (this.props.id_center !== 0) {
-      this.setState({ name: "" });
-      this.assoNameError.current.style.display = "none";
-      $("#modalAsso").modal("toggle");
+    if (this.props.id_asso !== 0) {
+      this.setState({ name: this.props.asso_name });
+      this.assoEditNameError.current.style.display = "none";
+      $("#modalAssoEdit").modal("toggle");
     } else {
       swal(
         "¡Atención!",
-        "Debe seleccionar un centro educativo de la lista.",
+        "Debe seleccionar una carrera asociada de la lista.",
         "warning"
       );
     }
@@ -50,35 +46,30 @@ export default class ModalAsso extends Component {
 
   /**
    * * Función que maneja el envío del formulario.
-   * * Se encarga de crear la carrera asociada si no se presentan
+   * * Se encarga de editar la carrera asociada si no se presentan
    * * errores en el nombre ingresado.
    */
   async handleSubmit(event) {
     event.preventDefault();
     const nameError = Validator.validateSimpleText(
       this.state.name,
-      this.assoNameError.current,
+      this.assoEditNameError.current,
       40,
       "textSpecial"
     );
     if (!nameError) {
       const assocareer = {
         name: this.state.name,
-        id_center: this.props.id_center,
       };
-      await axios.post(`/associated_career`, assocareer);
+      await axios.put(`/associated_career/` + this.props.id_asso, assocareer);
       this.props.getAssociatedCareer(this.props.id_center);
+      $("#modalAssoEdit").modal("hide");
+      swal("¡Listo!", "Se editó la carrera asociada exitosamente.", "success");
 
-      // * Dependiendo si vengo del modal, debo actualizar el select de mi grandparent
-      if (this.props.has_grand_parent) {
-        this.props.getAssociated();
-      }
-      $("#modalAsso").modal("hide");
-      swal(
-        "¡Listo!",
-        "Se creó la nueva carrera asociada exitosamente.",
-        "success"
-      );
+      this.props.refreshThis({
+        id_asso: 0,
+        asso_career_key: this.props.select_key + 1,
+      });
     }
   }
 
@@ -88,26 +79,27 @@ export default class ModalAsso extends Component {
         <button
           type="button"
           className="btn btn-primary btn-md"
-          data-target="#modalAsso"
+          data-target="#modalAssoEdit"
           onClick={this.validateShow}
         >
-          <i className="fas fa-plus"></i>
+          <i className="fas fa-edit"></i>
         </button>
-        <div className="modal fade" id="modalAsso" role="dialog">
+        <div className="modal fade" id="modalAssoEdit" role="dialog">
           <div className="modal-dialog modal-md modal-dialog-centered">
             <div className="modal-content">
               <div className="modal-header">
-                <h4 className="modal-title">Crear nueva carrera asociada</h4>
+                <h4 className="modal-title">Editar carrera asociada</h4>
                 <button type="button" className="close" data-dismiss="modal">
                   &times;
                 </button>
               </div>
               <div className="modal-body">
-                <p>Escriba el nombre de la carrera</p>
+                <p>Escriba el nuevo nombre de la carrera</p>
                 <div className="form-group">
                   <input
                     className="form-control"
                     type="text"
+                    id="nombreAsso"
                     name="name"
                     value={this.state.name}
                     onChange={this.handleChange}
@@ -115,7 +107,7 @@ export default class ModalAsso extends Component {
                   <div
                     className="alert alert-danger"
                     style={{ fontSize: 12 }}
-                    ref={this.assoNameError}
+                    ref={this.assoEditNameError}
                   ></div>
                 </div>
               </div>
@@ -124,7 +116,7 @@ export default class ModalAsso extends Component {
                   Cancelar
                 </button>
                 <button className="btn btn-primary" onClick={this.handleSubmit}>
-                  Crear
+                  Guardar
                 </button>
               </div>
             </div>
