+++ src/components/Modal/EditCenter.js
@@ -7,9 +7,9 @@ import { handleSimpleInputChange } from "../../helpers/Handles";
 
 /**
  * * Componente que muestra la ventana y elementos correspondientes
- * * para la creación de un nuevo centro educativo
+ * * para la edición de un centro educativo
  */
-export default class ModalCentro extends Component {
+export default class EditCenter extends Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -17,7 +17,7 @@ export default class ModalCentro extends Component {
     };
 
     //bind
-    this.show = this.show.bind(this);
+    this.validateShow = this.validateShow.bind(this);
     this.handleSubmit = this.handleSubmit.bind(this);
     this.handleChange = handleSimpleInputChange.bind(this);
 
@@ -26,15 +26,29 @@ export default class ModalCentro extends Component {
   }
 
   /**
-   * * Función que muestra el componente y limpia las variables
-   * * del estado, así como los mensajes de error correspondientes
+   * * Función que valida si el componente debe mostrarse, dependiendo
+   * * de las propiedades que le entran por parámetro. En este caso el
+   * * id del centro educativo debe estar definido.
    */
-  show() {
-    this.setState({ name: "" });
-    this.centerNameError.current.style.display = "none";
-    $("#modalCentro").modal("toggle");
+  validateShow() {
+    if (this.props.id_center !== 0) {
+      this.setState({ name: this.props.center_name });
+      this.centerNameError.current.style.display = "none";
+      $("#modalCentroEdit").modal("toggle");
+    } else {
+      swal(
+        "¡Atención!",
+        "Debe seleccionar un centro educativo de la lista.",
+        "warning"
+      );
+    }
   }
 
+  /**
+   * * Función que maneja el envío del formulario.
+   * * Se encarga de editar el centro educativo si
+   * * no se presentan errores en el nombre ingresado.
+   */
   async handleSubmit(event) {
     event.preventDefault();
     const nameError = Validator.validateSimpleText(
@@ -43,18 +57,20 @@ export default class ModalCentro extends Component {
       40,
       "textSpecial"
     );
+
     if (!nameError) {
       const center = {
         name: this.state.name,
       };
-      await axios.post(`/center`, center);
+      await axios.put(`/center/` + this.props.id_center, center);
       this.props.getCenter();
-      $("#modalCentro").modal("hide");
-      swal(
-        "¡Listo!",
-        "Se creó el nuevo centro educativo exitosamente.",
-        "success"
-      );
+      $("#modalCentroEdit").modal("hide");
+      swal("¡Listo!", "Se editó el centro exitosamente.", "success");
+      this.props.refreshThis({
+        id_center: 0,
+        center_key: this.props.select_key + 1,
+        associated_careers: [],
+      });
     }
   }
 
@@ -64,22 +80,22 @@ export default class ModalCentro extends Component {
         <button
           type="button"
           className="btn btn-primary btn-md"
-          data-target="#modalCentro"
-          onClick={this.show}
+          data-target="#modalCentroEdit"
+          onClick={this.validateShow}
         >
-          <i className="fas fa-plus"></i>
+          <i className="fas fa-edit"></i>
         </button>
-        <div className="modal fade" id="modalCentro" role="dialog">
+        <div className="modal fade" id="modalCentroEdit" role="dialog">
           <div className="modal-dialog modal-md modal-dialog-centered">
             <div className="modal-content">
               <div className="modal-header">
-                <h4 className="modal-title">Crear nuevo centro educativo</h4>
+                <h4 className="modal-title">Editar centro educativo</h4>
                 <button type="button" className="close" data-dismiss="modal">
                   &times;
                 </button>
               </div>
               <div className="modal-body">
-                <p>Escriba el nombre del centro</p>
+                <p>Escriba el nuevo nombre del centro</p>
                 <div className="form-group">
                   <input
                     className="form-control"
@@ -100,7 +116,7 @@ export default class ModalCentro extends Component {
                   Cancelar
                 </button>
                 <button className="btn btn-primary" onClick={this.handleSubmit}>
-                  Crear
+                  Guardar
                 </button>
               </div>
             </div>
