+++ src/services/env.js
@@ -1,0 +1,1 @@
+export const API = "http://localhost:4000";
